Fixed the issue where a line that's too long to be displayed in the terminal won't show.

+++ main/interface/cpp-terminal/platforms/platform.hpp
@@ -9,5 +9,5 @@ namespace Term::Private {
     // if the console is not connected
     std::pair<std::size_t, std::size_t> get_term_size();
 
-    unsigned int c32_display_width(const char32_t& c32);
+    short c32_display_width(const char32_t& c32);
 } // namespace Term
